Add the synthetic lidar generation and lidar data collection.
The python class Field contains a cartesian coordinate model
of the field and can, on demand, create a lidar view of the
data.

field.py contains enough helper classes to represent the robot
position and heading, as well as a simple command line driver
to move the robot around and view the synthetic lidar data that
represents that view.

lidar.py contains the class and driver to collect lidar data
from the serial port, sanitize the data, and output the lidar
data in a more digestible format.

The Field and LidarView classes are likely viable to do some
extensive automatic testing of the target detection algorithm.

TODO TASKS
Field   Upgrade it so it has about
500 points representing the actual profile of the back wall.  The
model in this commit has an overly simplified model.   The upgraded
version should be based on the precise published field dimensions.
The points beyond any of the aperatures in the back wall should be
located far enough away to look like infinity.  (> 600 inches)

Once the command line version of this code has been verified, the
test driver can be written.   Field model should have a method
that computes the model robot heading and distance from target.
(that's the center of the left goal)   The test driver needs to
be able to generate the lidar packet, ask the brain for a result,
and then compare the result with the result from the model.

This Lidar class needs to be upgraded/modified so that it outputs
data in the same format as the synthetic data.    Here is the
suggested output format:  (text)

888,PointCount
Degree[0],Distance[0]
Degree[1],Distance[1]
...
Degree[PointCount-1],Distance[PointCount-1]
999,Rh,(Rx, Ry)

888 marks the beginning of a single lidar rotation of data.
PointCount is the number of data points reported in the rotation.
This could be employed by the user of the data for buffer
allocation or just sanity checking.

999 marks the end of the rotation data.
The Rx, Ry, and Rh are the position and heading of the Robot (known
for synthetic data, these are likely zero coming from real data)

This particular format is chosen so that client can always expect
to read two numbers separate by a comma.   The 999 can trigger
additional reading if it is required.  (during self-test)

The lidar.py implementation provides detailed parsing of the packet,
but has broken checksum checking.    It still needs to be changed
to output the format documented above.

BRAIN will need to updated to read data from this stream of data
in the standard format.