Added plot_lidar script that plots files of lidar data.
Plots the contents of all files passed on the command line
one at a time.    Plots the data from 360 point polar data
form.

e.g.

0,24.5
1,36.9
...
359,777.0

Can be used to look at snapshots and for debugging.   The
output file for this can be produced from a single
collected lidar rotation data using the snippet:

for p in rotation:
    p.put_data