Cleanup.  Fix for serial readinto not working.  Fake data for no serial.
Some code restructuring.    Moved more from main into a few functions.
Moved around some comments associated with the packet format.

Changed the serial reading so it used read() instead of the readinto,
since that was not working for some reason.  Works now.

If there is no serial port open, then fake data is generated.   This
is useful for debugging changes to the full data collection path when
no lidar unit is available.   I think ultimately, I could plug in
the synthetic data generator and have it bounce around on the field
to get some interesting data.   But for now, it just replicates a single
packet for each slice of the rotation.