Added robot icon to plot, messed with rotation, and nixed backback wall.
Draw a blue robot icon in the center of the plot to provide info
on the robot orientation.  (to the right)

Messed with the orientation logic.  There is still some issue
with drawing after movement and rotation.  I have not debugged it
completely yet.

I removed the data for the backback wall.  It is showing up with
too much phantom data for now (seeing between the real data points)