Major refactoring of the Field code so it can be used as a data source.
This commit contains the all-important fix to the Robot model movement
so that it actually moves in the proper direction.   It was lacking
a degree to radians conversion in the move method.

This commit also contains a refactoring of the Field so that it can
work as a plug-in replacement for a lidar unit.   It produces a
FakeRotation, which is a compatible class with the lidar rotation.

Also I upgraded the field model so that it has 10 times more points
by generating a 10x scale of the field and then scaling it down.

I moved the plotting and file output functions in the LidarViewer
class.   That permits the various Rotations to export their data by
using the services of that class instead of them having their own
export and view functions.

The end result of all of this is that just about every line of code
has been changed.     Good thing is has not be modified by others
yet.