This is the quick-hack version of the code to read the LDS 2.1.
I just copied and modified the relevant classes and did a bit of
generalization to get things to work.    It doesn't really import
all of the data, but I was mostly just trying to debug the servo
control, so only cared about the rpm immediately after the header.

But our LDS v 2.1 doesn't appear to work well enough for me to
get this debugged completely.   It still needs some work to read
in all of the data, but it is nearly there.   I can eventually
see this being a plug-in and we could have some way to auto-detect
what version we were running.

Still needs to be properly refactored so we have Rotation and Packet
as base class, and derived classes for the different versions as
as appropriate.    Each could provide the header detection and
packet cracking for a generic driver.