Modified field tool for side-by-side plotting, testing out vector wall algo.
Enhanced the lidar_viewer so it does side-by-side plotting of lidar
source and the analysis data.  Added capability for different
analyzer data to be pushed to the plot in different colors.

Have the wall finding algo coded up (approximate), but it is inline
in the field simulation program.     Could be part of analyzer.

Added avg_distance to see how close we are to things.   Might need
to know that in order to tune the cooefficients of the algos.

matthew